requirements.txt
streamlit>=1.30.0
yfinance>=0.2.35
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.18.0
scipy>=1.11.0
openpyxl>=3.1.0
reportlab>=4.0.0
pyyaml>=6.0.0
pytest>=7.4.0
